feat: Implement Bloc and ViewModel factories
This commit introduces factories for creating Blocs and ViewModels.
This centralizes the creation of these objects and improves testability
by allowing for easier mocking and dependency injection. The Bloc
factory provides instances of AuthBloc, FavoritesBloc, ProfileBloc,
PlacesBloc, ThemeBloc, and LanguageBloc. The ViewModel factory
provides instances of LoginViewModel and SignUpViewModel. The
AppBlocProvider is also implemented to initialize and provide blocs.